requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
transformers>=4.30.0
torch>=2.0.0
prometheus-client>=0.17.0
locust>=2.15.0
Requests>=2.31.0
aiohttp>=3.9.0
grpcio>=1.54.0
grpcio-tools>=1.54.0
psutil>=5.9.0

